requirements.txt
discord.py==2.1.1
Flask==2.0.1
